Type NTXEntLoss and tests
